+++ sketch_oxide/src/range_filters/mod.rs
@@ -60,6 +60,7 @@ mod grafite;
 mod grf;
 mod memento_filter;
 mod pgm_index;
+mod rosetta;
 mod surf;
 
 pub use diva::DivaFilter;
@@ -67,6 +68,7 @@ pub use grafite::{Grafite, GrafiteStats};
 pub use grf::{GRFStats, GRF};
 pub use memento_filter::{MementoFilter, MementoStats};
 pub use pgm_index::PgmIndex;
+pub use rosetta::Rosetta;
 pub use surf::Surf;
 
 #[cfg(test)]